Fix error_reporting config for 32 bit systems
On PHP >= 8.0 (?? unverified, might be >=7.2 as well), when on a 32 bit
PHP installation, 0xffffffff is too big for a signed integer. Therefore,
it is casted to a float, which then dies due to a `TypeError` when passed
to `error_reporting()`, which expects `int`.

Either way, `0xffffffff` is a bad choice here:

* it relies on undocumented/internal behavior
* a named constant is always nicer