Set *last updated* date of a site to the *link date* on link
Leaving the `regat_lastdate` column undefined on INSERT errors with
MySql 8 due to the column not having a default value.

Since the date logic is custom, and might not work with NULL, and I
don't have any idea how to do DB migrations in this, the simplest
solution is to set it to the same value as the link date. That's not
logically wrong either.