Wechall: Update max score for HTS
A lot of users have > 100% reported.

It seems that the "onsitescore" variable, which was returned as
"maxscore" was way too low. Calculated a current potential max score
from https://www.hackthissite.org/user/rankings/explanation and used
this.

Updated the chall count, too.

Thanks Munto for pointing to the ranking explanation page.